+++ CHANGES.rst
@@ -68,6 +68,8 @@ Specviz2d
 Other Changes and Additions
 ---------------------------
 
+- Scatter viewer support in the deconfigged app instance. [#3778]
+
 4.4.2 (unreleased)
 ==================